Update README.md
This repo contains projects for Udacity Data Scientist Nanodegree. In this project, I have used SQL to explore a database related to movie rentals. I ran SQL queries and 
proposed some interesting questions about the database. The visualizations shows the output of the queries when I transfered the query results to excel.